reduce udp code to minimum
