--- src/main/udp_connector/handleUdpMessage.ts
@@ -1,5 +1,0 @@
-export default function createHandleMessage() {
-  return (newMessage: Buffer) => {
-    console.log(newMessage);
-  };
-}
